chore: Removed concurrency
After a discussion with Venelin Bakalov <venelin.bakalov@broadcom.com>
it was decided this is not needed.

This makes the build run concurrently instead of waiting on one another.

It may interfere with caching and it may be so that the jobs are writing
to the same cache (initially), but it's still worth it overall.

Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>